Revert "Revert "Dybdahl/refactor json reader""
This reverts commit b7c645f96271f4f27415af6757fef18499c63544.
