+++ document/src/main/java/com/yahoo/document/json/JsonWriter.java
@@ -38,6 +38,8 @@ import java.util.Iterator;
 import java.util.Map;
 
 import static com.yahoo.document.json.JsonSerializationHelper.*;
+import static com.yahoo.document.json.document.DocumentParser.FIELDS;
+import static com.yahoo.document.json.document.DocumentParser.REMOVE;
 
 /**
  * Serialize Document and other FieldValue instances as JSON.
@@ -106,7 +108,7 @@ public class JsonWriter implements DocumentWriter {
             // this makes it impossible to refeed directly, not sure what's correct
             // perhaps just change to "put"?
             generator.writeStringField("id", value.getId().toString());
-            generator.writeObjectFieldStart(JsonReader.FIELDS);
+            generator.writeObjectFieldStart(FIELDS);
 
             Iterator<Map.Entry<Field, FieldValue>> i = value.iterator();
             while (i.hasNext()) {
@@ -251,7 +253,7 @@ public class JsonWriter implements DocumentWriter {
         try {
             JsonGenerator throwAway = jsonFactory.createGenerator(out);
             throwAway.writeStartObject();
-            throwAway.writeStringField(JsonReader.REMOVE, docId.toString());
+            throwAway.writeStringField(REMOVE, docId.toString());
             throwAway.writeEndObject();
             throwAway.close();
         } catch (IOException e) {
