+++ document/src/test/java/com/yahoo/document/json/JsonReaderTestCase.java
@@ -30,6 +30,8 @@ import com.yahoo.document.datatypes.StringFieldValue;
 import com.yahoo.document.datatypes.Struct;
 import com.yahoo.document.datatypes.TensorFieldValue;
 import com.yahoo.document.datatypes.WeightedSet;
+import com.yahoo.document.json.document.DocumentParser;
+import com.yahoo.document.json.readers.DocumentParseInfo;
 import com.yahoo.document.update.AddValueUpdate;
 import com.yahoo.document.update.ArithmeticValueUpdate;
 import com.yahoo.document.update.ArithmeticValueUpdate.Operator;
@@ -63,6 +65,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import static com.yahoo.document.json.readers.SingleValueReader.*;
 import static com.yahoo.test.json.JsonTestHelper.inputJson;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
@@ -170,7 +173,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        DocumentPut put = (DocumentPut) r.readSingleDocument(JsonReader.SupportedOperation.PUT, "id:unittest:smoke::whee");
+        DocumentPut put = (DocumentPut) r.readSingleDocument(DocumentParser.SupportedOperation.PUT, "id:unittest:smoke::whee");
         smokeTestDoc(put.getDocument());
     }
 
@@ -181,7 +184,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": {"
                         + " \"assign\": \"orOther\" }}" + " }"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        DocumentUpdate doc = (DocumentUpdate) r.readSingleDocument(JsonReader.SupportedOperation.UPDATE, "id:unittest:smoke::whee");
+        DocumentUpdate doc = (DocumentUpdate) r.readSingleDocument(DocumentParser.SupportedOperation.UPDATE, "id:unittest:smoke::whee");
         FieldUpdate f = doc.getFieldUpdate("something");
         assertEquals(1, f.size());
         assertTrue(f.getValueUpdate(0) instanceof AssignValueUpdate);
@@ -194,7 +197,7 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"int1\": {"
                         + " \"assign\": null }}" + " }"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        DocumentUpdate doc = (DocumentUpdate) r.readSingleDocument(JsonReader.SupportedOperation.UPDATE, "id:unittest:smoke::whee");
+        DocumentUpdate doc = (DocumentUpdate) r.readSingleDocument(DocumentParser.SupportedOperation.UPDATE, "id:unittest:smoke::whee");
         FieldUpdate f = doc.getFieldUpdate("int1");
         assertEquals(1, f.size());
         assertTrue(f.getValueUpdate(0) instanceof ClearValueUpdate);
@@ -209,10 +212,10 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         smokeTestDoc(put.getDocument());
     }
 
@@ -225,10 +228,10 @@ public class JsonReaderTestCase {
                         + "\"put\": \"id:unittest:smoke::whee\""
                         + "}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         smokeTestDoc(put.getDocument());
     }
 
@@ -239,10 +242,10 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:smoke::whee\","
                         + " \"fields\": {}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         assertEquals("id:unittest:smoke::whee", parseInfo.documentId.toString());
     }
 
@@ -255,10 +258,10 @@ public class JsonReaderTestCase {
                         + "\"sandra\": \"person\","
                         + " \"cloud\": \"another person\"}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("skuggsjaa"));
         assertSame(Struct.class, f.getClass());
@@ -275,10 +278,10 @@ public class JsonReaderTestCase {
                         + " \"person\","
                         + " \"another person\"]}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         checkSimpleArrayAdd(doc);
     }
 
@@ -291,10 +294,10 @@ public class JsonReaderTestCase {
                         + " \"person\": 37,"
                         + " \"another person\": 41}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         Map<String, Integer> weights = new HashMap<>();
         FieldUpdate x = doc.getFieldUpdate("actualset");
         for (ValueUpdate<?> v : x.getValueUpdates()) {
@@ -320,11 +323,11 @@ public class JsonReaderTestCase {
                         + " \"element\": \"person\","
                         + " \"increment\": 13}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         Map<String, Tuple2<Number, String>> matches = new HashMap<>();
         FieldUpdate x = doc.getFieldUpdate("actualset");
         for (ValueUpdate<?> v : x.getValueUpdates()) {
@@ -346,13 +349,13 @@ public class JsonReaderTestCase {
     @Test
     public final void testArithmeticOperators() {
         Tuple2[] operations = new Tuple2[] {
-                new Tuple2<String, Operator>(JsonReader.UPDATE_DECREMENT,
+                new Tuple2<String, Operator>(UPDATE_DECREMENT,
                         ArithmeticValueUpdate.Operator.SUB),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_DIVIDE,
+                new Tuple2<String, Operator>(UPDATE_DIVIDE,
                         ArithmeticValueUpdate.Operator.DIV),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_INCREMENT,
+                new Tuple2<String, Operator>(UPDATE_INCREMENT,
                         ArithmeticValueUpdate.Operator.ADD),
-                new Tuple2<String, Operator>(JsonReader.UPDATE_MULTIPLY,
+                new Tuple2<String, Operator>(UPDATE_MULTIPLY,
                         ArithmeticValueUpdate.Operator.MUL) };
         for (Tuple2<String, Operator> operator : operations) {
             InputStream rawDoc = new ByteArrayInputStream(
@@ -362,11 +365,11 @@ public class JsonReaderTestCase {
                             + " \"" + (String) operator.first + "\": 13}}}}"));
 
             JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-            JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+            DocumentParseInfo parseInfo = r.parseDocument().get();
             DocumentType docType = r.readDocumentType(parseInfo.documentId);
             DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
-            r.readUpdate(doc);
+            r.readUpdate(parseInfo.fieldsBuffer, doc);
             Map<String, Tuple2<Number, Operator>> matches = new HashMap<>();
             FieldUpdate x = doc.getFieldUpdate("actualset");
             for (ValueUpdate v : x.getValueUpdates()) {
@@ -396,11 +399,11 @@ public class JsonReaderTestCase {
                         + " \"element\": 3,"
                         + " \"assign\": \"nalle\"}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
 
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         Map<Number, String> matches = new HashMap<>();
         FieldUpdate x = doc.getFieldUpdate("actualarray");
         for (ValueUpdate v : x.getValueUpdates()) {
@@ -434,10 +437,10 @@ public class JsonReaderTestCase {
                         + " \"nalle\": 2,"
                         + " \"tralle\": 7 }}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("actualset"));
         assertSame(WeightedSet.class, f.getClass());
@@ -455,10 +458,10 @@ public class JsonReaderTestCase {
                         + " \"nalle\","
                         + " \"tralle\"]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("actualarray"));
         assertSame(Array.class, f.getClass());
@@ -476,10 +479,10 @@ public class JsonReaderTestCase {
                         + " { \"key\": \"nalle\", \"value\": \"kalle\"},"
                         + " { \"key\": \"tralle\", \"value\": \"skalle\"} ]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("actualmap"));
         assertSame(MapFieldValue.class, f.getClass());
@@ -495,10 +498,10 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:testsinglepos::bamf\","
                         + " \"fields\": { \"singlepos\": \"N63.429722;E10.393333\" }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("singlepos"));
         assertSame(Struct.class, f.getClass());
@@ -512,10 +515,10 @@ public class JsonReaderTestCase {
                 Utf8.toBytes("{\"put\": \"id:unittest:testsinglepos::bamf\","
                         + " \"fields\": { \"singlepos\": \"W46.63;S23.55\" }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("singlepos"));
         assertSame(Struct.class, f.getClass());
@@ -533,10 +536,10 @@ public class JsonReaderTestCase {
                         + "\""
                         + " }}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue(doc.getField("actualraw"));
         assertSame(Raw.class, f.getClass());
@@ -553,10 +556,10 @@ public class JsonReaderTestCase {
                         + "{ \"key\": \"bamse\", \"value\": [1, 2, 3] }"
                         + "]}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         Document doc = put.getDocument();
         FieldValue f = doc.getFieldValue("actualMapStringToArrayOfInt");
         assertSame(MapFieldValue.class, f.getClass());
@@ -575,10 +578,10 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": {"
                         + " \"assign\": \"orOther\" }}" + " }"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         FieldUpdate f = doc.getFieldUpdate("something");
         assertEquals(1, f.size());
         AssignValueUpdate a = (AssignValueUpdate) f.getValueUpdate(0);
@@ -594,10 +597,10 @@ public class JsonReaderTestCase {
                         + "{ \"key\": \"bamse\", \"value\": [1, 2, 3] }"
                         + "]}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         FieldUpdate f = doc.getFieldUpdate("actualMapStringToArrayOfInt");
         assertEquals(1, f.size());
         AssignValueUpdate assign = (AssignValueUpdate) f.getValueUpdate(0);
@@ -618,10 +621,10 @@ public class JsonReaderTestCase {
                         + " \"person\": 37,"
                         + " \"another person\": 41}}}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentUpdate doc = new DocumentUpdate(docType, parseInfo.documentId);
-        r.readUpdate(doc);
+        r.readUpdate(parseInfo.fieldsBuffer, doc);
         FieldUpdate x = doc.getFieldUpdate("actualset");
         assertEquals(1, x.size());
         AssignValueUpdate assign = (AssignValueUpdate) x.getValueUpdate(0);
@@ -819,12 +822,12 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"smething\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
         exception.expect(NullPointerException.class);
         exception.expectMessage("Could not get field \"smething\" in the structure of type \"smoke\".");
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
     }
 
     @Test
@@ -848,10 +851,10 @@ public class JsonReaderTestCase {
                         + " \"fields\": { \"something\": \"smoketest\","
                         + " \"nalle\": \"bamse\"}}"));
         JsonReader r = new JsonReader(types, rawDoc, parserFactory);
-        JsonReader.DocumentParseInfo parseInfo = r.parseDocument().get();
+        DocumentParseInfo parseInfo = r.parseDocument().get();
         DocumentType docType = r.readDocumentType(parseInfo.documentId);
         DocumentPut put = new DocumentPut(new Document(docType, parseInfo.documentId));
-        r.readPut(put);
+        r.readPut(parseInfo.fieldsBuffer, put);
         smokeTestDoc(put.getDocument());
     }
